+++ pkg/glob/CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 1.0.0+1
 
 * Fix several analyzer warnings.
+
+* Fix the tests on Windows.